Stop Cider from asking us whether it should open our browser
Since we're serving the static assets via wrangler -- to bypass building out CORS support in the
backend -- we don't connect to the URL that Cider gives us.

The better dev experience would be to add CORS support to the backend so it matches the normal Cider
experience. Also, CORS is "good", in the general way that it's security best practice, but not in
the way that I can speak definitively about its pros and cons. One day it'll get added, if it's
necessary and good.