Persist game data and cart values on cart changes
I don't care to persist the games if there aren't any changes to the cart, but when there are
changes to the cart, I'd like the cart -- and by extension the games -- to be saved.

There's a more nuanced way to reduce the amount of data we need to save, but that's going to be
future work.