Revert "Switch UI creation to new CreateRoot React api"
This reverts commit 2b48f9e1a6c8ba6f838451a3d16a1ab4dd5dc922.

The subscription for the current page, on the root component, was getting disposed of, though I
don't know the trigger for this. The result is that navigation stopped working, because the root
component wouldn't get the updated `:current-page` value, and thus wouldn't switch what it was
rendering.

Since this was new behavior, and going on my intuition, I suspect this is caused by some oddity
around the new CreateRoot behavior in React 18, and thus I'm reverting that change to see if this
new nav issue goes away.