Switch build system over to asdf
`jenv` was nice at first, in that it made it super simple to specify which version of java I wanted
to use, but once I upgraded my version of java, I also had to change my invocation of jenv to
register the new version.

That was a lot of tedium, so instead I've decided to manage my java versions with ASDF, especially
because it can manage Node versions as well.

Furthermore, I believe the build environment in Cloudflare has ASDF installed, so I'm going to try
switching the build script over to making use of it, and see what happens.