Move ASDF plugin installation into Cloudflare ENV vars
We have to use the `UNSTABLE_PRE_BUILD` env var in our build process to get the asdf `java` plugin
installed before the build system installs the tool versions defined in our `.tool-versions` file.

I'm wary of using an ENV var labeled "UNSTABLE", as it might disappear without warning, but also
this project has effectively no risk, so I'm going to go with the ease of use it affords us.